+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v24.20
 
 ### Pre-releases
+- `v24.20-alpha11`
+- `v24.20-alpha10`
+- `v24.20-alpha9`
 - `v24.20-alpha8`
 - `v24.20-alpha7`
 - `v24.20-alpha6`
@@ -13,19 +16,27 @@
 - `v24.20-alpha1`
 
 ### Breaking changes
+- External login endpoints changed (#384, `v24.20-alpha9`)
 - Default password criteria are more restrictive (#372, `v24.20-alpha1`, Compatible with Seacat Auth Webui v24.19-alpha and later, Seacat Account Webui v24.08-beta and later)
 
 ### Fix
+- Fix AttributeError in credentials update (#399, `v24.20-alpha11`)
+- Catch token decoding errors when finding sessions (#397, `v24.20-alpha10`)
+- Properly encrypt cookie value in session update (#394, `v24.20-alpha8`)
+- Properly parse URL query before adding new parameters (#393, `v24.20-alpha8`)
 - Delete client cookie on introspection failure (#385, `v24.20-alpha6`)
 - Extend session expiration at cookie entrypoint (#383, `v24.20-alpha5`)
 - Do not log failed LDAP login as error (#381, `v24.20-alpha4`)
 - Properly handle Argon2 verification error in login call (#378, `v24.20-alpha3`)
 
 ### Features
+- External login with dynamic redirection (#384, `v24.20-alpha9`)
+- Custom response codes for tenant-related authorization errors (#392, `v24.20-alpha8`)
 - Implement OAuth refresh tokens (#358, `v24.20-alpha2`)
 - Configurable password criteria (#372, `v24.20-alpha1`)
 
 ### Refactoring
+- Expected client ID is optional for authentication (#390, `v24.20-alpha8`)
 - Move SSO session creation into session service (#387, `v24.20-alpha8`)
 - Refactor duplicate code in session build (#386, `v24.20-alpha7`)